+++ README.md
@@ -3,3 +3,7 @@
 ```bash
 go run github.com/romnnn/mongoimport/cmd/mongoimport csv <path-to-csv>
 ```
+
+```bash
+go run github.com/romnnn/mongoimport/examples/multi
+```
