+++ examples/data/ford_escort2.csv
@@ -1,0 +1,12 @@
+"Year", "Mileage (thousands)",  "Price"
+1997,  24,    9990 
+1997,  37,    9990 
+1997,  29,    9990 
+1997,  70,    8990 
+1996,  29,    7990 
+1995,  72,    5994 
+1993,  72,    5994 
+1994,  61,    5500 
+1998,   7,   11000 
+1998,   12,   11010
+1999,   25,   1020
